lmb-1053: make a constant for the URI of besluit and Artikel
