+++ config/ldes-delta-pusher/handle-types-util.ts
@@ -2,6 +2,9 @@ import { Changeset } from "../types";
 import { log } from "./logger";
 import { InterestingSubject, publish } from "./publisher";
 
+export const ARTIKEL_TYPE_URI = "http://data.vlaanderen.be/ns/besluit#Artikel";
+export const BESLUIT_TYPE_URI = "http://data.vlaanderen.be/ns/besluit#Besluit";
+
 type SubjectFilter = (subjects: string[]) => Promise<InterestingSubject[]>;
 type SubjectAddition = (subject: InterestingSubject) => Promise<string>;
 
@@ -25,7 +28,7 @@ const filterInterestingSubjects = async (
   let subjectsToKeep: InterestingSubject[] = [];
 
   const chunkSize = 1000;
-  for (let i = 0; i < allSubjects.length; i += chunkSize) {
+  for (let i = 0;i < allSubjects.length;i += chunkSize) {
     const currentChunk = allSubjects.slice(i, i + chunkSize);
     const toKeepForChunk = await filter(currentChunk);
     subjectsToKeep = subjectsToKeep.concat(toKeepForChunk);
