+++ config/ldes-delta-pusher/initialization.ts
@@ -1,6 +1,8 @@
+import { ARTIKEL_TYPE_URI, BESLUIT_TYPE_URI } from './handle-types-util';
+
 export const initialization = {
   public: {
-    "http://data.vlaanderen.be/ns/besluit#Artikel": {},
-    "http://data.vlaanderen.be/ns/besluit#Besluit": {},
+    ARTIKEL_TYPE_URI: {},
+    BESLUIT_TYPE_URI: {},
   },
 };
